+++ src/main/java/pro/sky/finalprojectsky/repository/ImageRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.finalprojectsky.entity.Image;
 
 public interface ImageRepository extends JpaRepository<Image, Integer> {
-    Image findByAdsId(long adsId);
-
+    Image findByAdsId(Integer adsId);
 }
